+++ company/tests/test_views.py
@@ -2019,7 +2019,7 @@ def test_verify_companies_house_good_access_token(
 def test_create_transfer_ownership_invite(authed_client, authed_supplier):
 
     data = {'new_owner_email': 'foo@bar.com'}
-    url = reverse('transfer-ownership-invite')
+    url = reverse('old-transfer-ownership-invite')
 
     response = authed_client.post(url, data=data)
 
@@ -2058,7 +2058,7 @@ def test_create_duplicated_transfer_ownership_invite(
         'company': authed_supplier.company.pk,
     }
     response = authed_client.post(
-        reverse('transfer-ownership-invite'),
+        reverse('old-transfer-ownership-invite'),
         data=data,
         format='json'
     )
@@ -2133,7 +2133,7 @@ def test_retrieve_transfer_ownership_invite(
     invite.save()
 
     response = authed_client.get(
-        reverse('transfer-ownership-invite-detail',
+        reverse('old-transfer-ownership-invite-detail',
                 kwargs={'uuid': str(invite.uuid)})
     )
 
@@ -2166,7 +2166,7 @@ def test_accept_transfer_ownership_invite(
     )
     invite.save()
     response = authed_client.patch(
-        reverse('transfer-ownership-invite-detail',
+        reverse('old-transfer-ownership-invite-detail',
                 kwargs={'uuid': str(invite.uuid)}),
         {'accepted': True}
     )
@@ -2203,7 +2203,7 @@ def test_accept_transfer_ownership_invite_case_insensitive(
     )
     invite.save()
     response = authed_client.patch(
-        reverse('transfer-ownership-invite-detail',
+        reverse('old-transfer-ownership-invite-detail',
                 kwargs={'uuid': str(invite.uuid)}),
         {'accepted': True}
     )
@@ -2239,7 +2239,7 @@ def test_accept_wrong_transfer_ownership_invite(
     )
     invite.save()
     response = authed_client.patch(
-        reverse('transfer-ownership-invite-detail',
+        reverse('old-transfer-ownership-invite-detail',
                 kwargs={'uuid': str(invite.uuid)}),
         {'accepted': True}
     )
@@ -2278,7 +2278,7 @@ def test_accept_transfer_ownership_invite_to_collaborator(
         requestor=existing_owner,
     )
     response = authed_client.patch(
-        reverse('transfer-ownership-invite-detail',
+        reverse('old-transfer-ownership-invite-detail',
                 kwargs={'uuid': str(invite.uuid)}),
         {'accepted': True}
     )
@@ -2301,7 +2301,7 @@ def test_accept_transfer_ownership_invite_supplier_has_other_company(
     )
     invite.save()
     response = authed_client.patch(
-        reverse('transfer-ownership-invite-detail',
+        reverse('old-transfer-ownership-invite-detail',
                 kwargs={'uuid': str(invite.uuid)}),
         {'accepted': True}
     )
@@ -2334,7 +2334,7 @@ def test_accept_transfer_ownership_invite_requestor_not_legit(
     )
     invite.save()
     response = authed_client.patch(
-        reverse('transfer-ownership-invite-detail',
+        reverse('old-transfer-ownership-invite-detail',
                 kwargs={'uuid': str(invite.uuid)}),
         {'accepted': True}
     )
@@ -2356,7 +2356,7 @@ def test_company_create_collaboration_invite(
     authed_client, authed_supplier
 ):
     data = {'collaborator_email': 'foo@bar.com'}
-    url = reverse('collaboration-invite-create')
+    url = reverse('old-collaboration-invite-create')
     response = authed_client.post(url, data=data)
 
     invite = models.CollaboratorInvite.objects.get(
@@ -2387,7 +2387,7 @@ def test_company_create_duplicated_collaboration_invite(
     )
 
     data = {'collaborator_email': 'foo@bar.com'}
-    url = reverse('collaboration-invite-create')
+    url = reverse('old-collaboration-invite-create')
     response = authed_client.post(url, data=data)
 
     assert response.status_code == status.HTTP_400_BAD_REQUEST
@@ -2415,7 +2415,7 @@ def test_accept_collaborator_invite(
     )
 
     url = reverse(
-        'collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
+        'old-collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
     )
     response = authed_client.patch(url, {'accepted': True})
     assert response.status_code == 200
@@ -2449,7 +2449,7 @@ def test_accept_collaborator_invite_case_insensitive(
     )
 
     url = reverse(
-        'collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
+        'old-collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
     )
     response = authed_client.patch(url, {'accepted': True})
     assert response.status_code == 200
@@ -2484,7 +2484,7 @@ def test_accept_wrong_collaborator_invite(
     )
 
     url = reverse(
-        'collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
+        'old-collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
     )
     response = authed_client.patch(url, {'accepted': True})
     assert response.status_code == status.HTTP_400_BAD_REQUEST
@@ -2513,7 +2513,7 @@ def test_accept_collaborator_invite_supplier_has_other_company(
         requestor=authed_supplier,
     )
     url = reverse(
-        'collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
+        'old-collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
     )
     response = authed_client.patch(url, {'accepted': True})
     error = serializers.InviteSerializerMixin.MESSAGE_ALREADY_HAS_COMPANY
@@ -2544,7 +2544,7 @@ def test_accept_collaborator_invite_requestor_not_legit(
         requestor=supplier
     )
     url = reverse(
-        'collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
+        'old-collaboration-invite-detail', kwargs={'uuid': str(invite.uuid)}
     )
     response = authed_client.patch(url, {'accepted': True})
     error = serializers.InviteSerializerMixin.MESSAGE_INVALID_REQUESTOR
@@ -2562,9 +2562,9 @@ def test_accept_collaborator_invite_requestor_not_legit(
 
 @pytest.mark.django_db
 @pytest.mark.parametrize('url', (
-    reverse('collaboration-invite-create'),
+    reverse('old-collaboration-invite-create'),
     reverse('remove-collaborators'),
-    reverse('transfer-ownership-invite'),
+    reverse('old-transfer-ownership-invite'),
 ))
 def test_multi_user_account_management_views_forbidden(
     url, authed_client, authed_supplier