+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - TT-1733 - Added request identity verification feature
 - TT-1733 - Allow name to be provided on supplier create
 - TT-1716 - Allow disconnecting self from company
+- TT-1727 - Consolidate collaboration models in backwards compatible way.
 
 ### Fixed bugs
 - No ticket - Upgrade django to 1.11.23 to fix vulnerability